+++ src/main/java/com/aviad/coupons/dal/IUsersDal.java
@@ -21,7 +21,7 @@ public interface IUsersDal extends CrudRepository<UserEntity, Long> {
 
     @Query("SELECT new com.aviad.coupons.dto.User (u.id, u.username, u.email, u.userType, u.company.id)" +
             "FROM UserEntity u WHERE u.id = :id")
-    User getUser(@Param("id") long id);
+    User getUser(@Param("id") int id);
 
     @Query("SELECT new com.aviad.coupons.dto.User (u.id, u.username, u.email, u.userType, u.company.id)" +
             "FROM UserEntity u WHERE u.company.id = :id")
